requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
fastapi==0.94.1
pdfplumber==0.8.0
Pillow==9.4.0
PyMuPDF==1.21.1
pytesseract==0.3.10
torch==1.10.2+cpu
torchvision==0.11.3+cpu
transformers==4.27.1
uvicorn[standard]==0.21.0
easyocr==1.7.1
huggingface-hub==0.16.4
jinja2==3.1.2
numpy==1.25.2
opencv-python==4.8.0.76
pandas==2.1.0
pdf2image==1.16.3
pillow==10.0.0
pymupdf==1.22.5
pytesseract==0.3.10
scikit-image==0.21.0
tesseract==0.1.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.31.0
